# src/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Primera entrega del proyecto final del curso de Backend con Nodejs en CoderHouse.",
    "version": "0.1.0",
    "title": "e-commerce"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "schemes": ["http"],
...
  "paths": {
    "/products": {
      "post": {
        "tags": ["products"],
        "summary": "add product",
        "description": "add product",
        "operationId": "addProduct",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "product object that needs to be added",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product_variant1"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "get": {
        "tags": ["products"],
        "summary": "get all products",
        "description": "get all products",
        "operationId": "getAllProducts",
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": ["products"],
        "summary": "get a product by id",
        "description": "get a product by id",
        "operationId": "getProductById",
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "put": {
        "tags": ["products"],
        "summary": "update a product by id",
        "description": "update a product by id",
        "operationId": "updateProduct",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "product object that needs to be updated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product_variant1"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": ["products"],
        "summary": "delete a product by id",
        "description": "delete a product by id",
        "operationId": "deleteProduct",
        "produces": ["application/json"],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Product_variant1": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "Guitarra"
        },
        "description": {
          "type": "string",
          "example": "Guitarra Alpujarra"
        },
        "code": {
          "type": "string",
          "example": "6 cuerdas"
        },
        "thumbnail": {
          "type": "string",
          "example": "https://assets.stickpng.com/images/580b585b2edbce24c47b289d.png"
        },
        "stock": {
          "type": "number",
          "example": 20
        },
        "price": {
          "type": "number",
          "example": 950.5
        }
      }
    }
  }
}